requirements.txt
flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
selenium==4.18.1
exa-py==1.0.6
groq==0.4.2
langgraph==0.0.26
beautifulsoup4==4.12.3
python-dotenv==1.0.1
fake-useragent==1.4.0
urllib3==2.2.1
certifi==2024.2.2
pandas==2.1.0
twilio==8.3.0
apscheduler==3.10.1
selenium>=4.15.2
webdriver-manager>=4.0.1
requests-html==0.10.0
nltk==3.8.1
feedparser==6.0.11
gunicorn==21.2.0
undetected-chromedriver>=3.5.4
exa-py>=1.0.0
groq>=0.4.2
python-dateutil==2.9.0
gunicorn==21.2.0
logging-utils==1.0.2
schedule==1.2.1
webdriver-manager==4.0.1
requests-html==0.10.0
undetected-chromedriver==3.5.4
langgraph>=0.0.19
